pyproject.toml
[tool.poetry]
name = "pipeline-docker"
version = "0.1.0"
description = ""
authors = ["Paul Hetherington <ph@mystic.ai>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.71.0"
pipeline-ai = "^0.0.12"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
